Make it a little bit more generic
