+++ core-ui/src/components/Predefined/Create/index.js
@@ -1,2 +1,1 @@
-export * from './Namespace/Namespace.create.form.js';
-export * from './Namespace/Namespace.create.modal.js';
+export * from './Namespace/Namespace.create.js';
